requirements.txt
numpy
matplotlib
opencv-python
scipy
rpi-gpio --pre